overhaul project into python module
